detail_product

+++ Controllers/UserController.cs
@@ -8,6 +8,8 @@ using final_project.Models.Entities;
 using final_project.Controllers;
 using final_project;
 using Microsoft.AspNetCore.Authorization;
+using System.IdentityModel.Tokens.Jwt;
+
 namespace final_project.Controllers
 {  
     [Route("api/[controller]")]
@@ -21,32 +23,69 @@ namespace final_project.Controllers
             _context = context;
         }
         [HttpGet]
-        [Authorize(Roles = "Admin")]
+        [Authorize]
         public ActionResult<IEnumerable<User>> Get()
-        {
+        {  //Get all user chi co admin dc get
+            
+             var handler = new JwtSecurityTokenHandler();
+            string authHeader = Request.Headers["Authorization"];
+            authHeader = authHeader.Replace("Bearer ", "");
+            var jsonToken = handler.ReadToken(authHeader);
+             var tokenS = handler.ReadToken(authHeader) as JwtSecurityToken;
+
+        
+        var Role=tokenS.Claims.First(claim => claim.Type =="Role").Value;
+        if(Role=="Admin")
             return _userService.GetUsers();
+            else return BadRequest();
         }
 
 
         [HttpGet("{id}")]
         [Authorize]
         public ActionResult<User> Get(string id)
-        {   
-            return _userService.GetUserById(id);
+        {   //Get user by id chi co user do or admin dc get
+            var handler = new JwtSecurityTokenHandler();
+            string authHeader = Request.Headers["Authorization"];
+            authHeader = authHeader.Replace("Bearer ", "");
+            var jsonToken = handler.ReadToken(authHeader);
+             var tokenS = handler.ReadToken(authHeader) as JwtSecurityToken;
+
+        var ID = tokenS.Claims.First(claim => claim.Type =="ID").Value;
+        var Role=tokenS.Claims.First(claim => claim.Type =="Role").Value;
+        if(ID==id||Role=="Admin")
+        return _userService.GetUserById(id);
+        else return BadRequest();
         }
 
         [HttpPut("{id}")]
         [Authorize]
         public void Put(string id, [FromBody] User user)
-        {
+        {  //Chi co user do or admin dc update infor
+            var handler = new JwtSecurityTokenHandler();
+            string authHeader = Request.Headers["Authorization"];
+            authHeader = authHeader.Replace("Bearer ", "");
+            var jsonToken = handler.ReadToken(authHeader);
+             var tokenS = handler.ReadToken(authHeader) as JwtSecurityToken;
+
+        var ID = tokenS.Claims.First(claim => claim.Type =="ID").Value;
+        var Role=tokenS.Claims.First(claim => claim.Type =="Role").Value;
+        if(ID==id||Role=="Admin")
             _userService.UpdateUser(id,user);
         }
          
 
         [HttpDelete("{id}")]
-        [Authorize(Roles = "Admin")]
+        [Authorize]
         public void Delete(string id)
-        {
+        {  //Chi co admin moi dc quyen xoa
+             var handler = new JwtSecurityTokenHandler();
+            string authHeader = Request.Headers["Authorization"];
+            authHeader = authHeader.Replace("Bearer ", "");
+            var jsonToken = handler.ReadToken(authHeader);
+        var tokenS = handler.ReadToken(authHeader) as JwtSecurityToken;
+        var Role=tokenS.Claims.First(claim => claim.Type =="Role").Value;
+        if(Role=="Admin")
             _userService.DeleteUser(id);
         }
 
@@ -61,8 +100,17 @@ namespace final_project.Controllers
         [HttpGet("email={email}")]
         [Authorize]
         public ActionResult<User> GetUser(string email)
-        {   
+        {   //Get user by email chi co user do moi duoc quyen get
+            var handler = new JwtSecurityTokenHandler();
+            string authHeader = Request.Headers["Authorization"];
+            authHeader = authHeader.Replace("Bearer ", "");
+            var jsonToken = handler.ReadToken(authHeader);
+            var tokenS = handler.ReadToken(authHeader) as JwtSecurityToken;
+
+            var ID = tokenS.Claims.First(claim => claim.Type =="ID").Value;
+             if(_userService.GetUserByEmail(email).id==ID)
             return _userService.GetUserByEmail(email);
+            else return BadRequest();
         }
         
     }
